+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on Keep a Changelog (https://keepachangelog.com), and this p
 - Add missing unit tests for parser helper (PrintStmt).
 - Add missing unit tests for parser debug (DimStmt).
 - Add missing unit tests for parser helper (DimStmt).
+- Add new test example for flow control (`FOR ... TO ... STEP ... NEXT`).
 
 ## [Unreleased] - 2026-02-02
 ### Added